namespaceを除去
