+++ includes/event.h
@@ -3,7 +3,7 @@
 
 #include <stdexcept>
 #include <string>
-#include "message.h"
+#include "command.h"
 #include "error_status.h"
 
 class User;
@@ -14,10 +14,10 @@ class Event {
 		virtual ~Event();
 		int		get_fd(void) const;
 		int		get_event_type(void) const;
-		const message::Command&	get_command(void) const;
+		const Command&	get_command(void) const;
 		const std::vector<std::string>&	get_command_params(void) const;
 		const ErrorStatus& get_error_status(void) const;
-		void	set_command(const message::Command& command);
+		void	set_command(const Command& command);
 		void	set_command_params(const std::vector<std::string>& commmand_params);
 		void	set_error_status(const ErrorStatus&);
 		bool	HasErrorOccurred(void) const;
@@ -55,7 +55,7 @@ class Event {
 	private:
 		const int	fd_;
 		const int	event_type_;
-		const message::Command* command_;
+		const Command* command_;
 		std::vector<std::string>	command_params_;
 		const ErrorStatus* error_status_;
 		const User* executer_;
