+++ test/MessageErrorTest.cpp
@@ -1,17 +1,18 @@
 #include <gtest/gtest.h>
-#include "message.h"
+#include "message_parser.h"
+#include "command.h"
 
 //test not Found command
 TEST(MessageErrorTest, ParsingErrorTest1) {
 
   std::string raw_message =
     "PASS_ param1 param2 param3 \r\n";
-  message::MessageParser message_parser(raw_message);
+  MessageParser message_parser(raw_message);
 
   // Expect not found command.
-  EXPECT_EQ(message::kNotFound, message_parser.get_command());
+  EXPECT_EQ(kNotFound, message_parser.get_command());
   // Expect ParseError state.
-  EXPECT_EQ(message::kParseError, message_parser.get_state());
+  EXPECT_EQ(kParseError, message_parser.get_state());
 }
 
 //test empty command
@@ -19,12 +20,12 @@ TEST(MessageErrorTest, ParsingErrorTest2) {
 
   std::string raw_message =
     "\n";
-  message::MessageParser message_parser(raw_message);
+  MessageParser message_parser(raw_message);
 
   //Expect not found command.
-  EXPECT_EQ(message::kNotFound, message_parser.get_command());
+  EXPECT_EQ(kNotFound, message_parser.get_command());
   //Expect ParseError state.
-  EXPECT_EQ(message::kParseEmpty, message_parser.get_state());
+  EXPECT_EQ(kParseEmpty, message_parser.get_state());
 }
 
 //test empty command
@@ -32,12 +33,12 @@ TEST(MessageErrorTest, ParsingErrorTest3) {
 
   std::string raw_message =
     "\r\n";
-  message::MessageParser message_parser(raw_message);
+  MessageParser message_parser(raw_message);
 
   //Expect not found command.
-  EXPECT_EQ(message::kNotFound, message_parser.get_command());
+  EXPECT_EQ(kNotFound, message_parser.get_command());
   //Expect ParseError state.
-  EXPECT_EQ(message::kParseEmpty, message_parser.get_state());
+  EXPECT_EQ(kParseEmpty, message_parser.get_state());
 }
 
 //test not ascii code command
@@ -45,10 +46,10 @@ TEST(MessageErrorTest, ParsingErrorTest4) {
 
   std::string raw_message =
     "あいうえお param1 param2\r\n";
-  message::MessageParser message_parser(raw_message);
+  MessageParser message_parser(raw_message);
 
   //Expect not found command.
-  EXPECT_EQ(message::kNotFound, message_parser.get_command());
+  EXPECT_EQ(kNotFound, message_parser.get_command());
   //Expect ParseError state.
-  EXPECT_EQ(message::kParseNotAscii, message_parser.get_state());
+  EXPECT_EQ(kParseNotAscii, message_parser.get_state());
 }
